supplychain-backend/requirements.txt
psycopg2-binary==2.9.6
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
